amdxdna: aie2_smu: remove busy wait until SMU_RESP_REG before submit commands
As the aie2_smu_exec() function is the only funciton in XDNA driver to submit
SMU commands and wait until it has finished. And the function does lock around
the SMU registers access. And the NPU SMU is only used by xdna. Removed the
need to poll the SMU_RESP_REG until it is cleared before commiting new SMU command.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>