Build fixes. (#891)
* CMake/pkg: Don't fail on unsupported Linux distros.

Instead of emitting a FATAL_ERROR, emit a WARNING.
Besides this FATAL_ERROR the build succeeds.

This makes the build not to refuse to work on Arch Linux.

Signed-off-by: Lubosz Sarnecki <lubosz@gmail.com>

* CMake/pkg: Remove trailing whitespace.

Signed-off-by: Lubosz Sarnecki <lubosz@gmail.com>

* shim/cmake: Use CMake generator expression to determine build type.

Use the modern way of determining a debug build.
In addition, only set XDNA_SHIM_DEBUG on the XDNA_TARGET.

I tested this patch with all occurances of XDNA_SHIM_DEBUG in the code
and it behaves as expected.

This removes the old string comparison which failed when
CMAKE_BUILD_TYPE was not set.

Following CMake error was triggered in that case:
```
CMake Error at src/shim/CMakeLists.txt:39 (if):
  if given arguments:

    "STREQUAL" "Debug"

  Unknown arguments specified
```

Instead of just wrapping the string comparison with "", which would
resolve the issue, I have decided to instead implement the modern way of
detecting debug mode.

See:
https://cmake.org/cmake/help/latest/manual/cmake-generator-expressions.7.html#conditional-expressions
Signed-off-by: Lubosz Sarnecki <lubosz@gmail.com>

---------

Signed-off-by: Lubosz Sarnecki <lubosz@gmail.com>